% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcalc.R
\name{agregate_data}
\alias{agregate_data}
\title{Measure aggregation by formula}
\usage{
agregate_data(.data, f, id_cols = NULL, value_cols = NULL, na.rm = FALSE)
}
\arguments{
\item{.data}{Dataset (data.frame), 'id' column is required.}

\item{f}{A formula}

\item{id_cols}{Identification columns except id (string vector or R formula, like ~y)}

\item{value_cols}{Calculation columns (string vector or R formula, like ~y)}

\item{na.rm}{TRUE to replace all NAs to zero}
}
\value{
Calculation results (data.frame containing id_cols and value_cols)
        Results are placed to value_cols, id_cols copied without changes
}
\description{
Performs a measures aggregation by formula on a dataset
}
\details{
The function transforms a dataset so measure IDs become dataframe columns and then
applies a formula to calculate aggregated measures.

A formula can be a mesaure ID or any single-parameter function with
aggregation operators (+. -, etc). See examples below.

If a dataset contain additional identification variables, id_cols should be provided.

A value_cols parameter specifies which variables are to be aggregated. If omitted,
any numeric or logical columns are taken.
}
\examples{
\dontrun{

d <- data.frame(id = c("i1", "i1", "i2", "i3"), y = c(10, 20, 10, 10), a = c(4,2,3, 1), r = c(4,5,6,7))
agregate_data(d, f = "log(i1)+i2+i3", id_cols = ~y, value_cols = c("a", "r"), na.rm = TRUE)

d <- data.frame(id = c("i1", "i2", "i3"), a = c(4, 3, 1), r = c(4, 6,7))
agregate_data(d, f = "i1+i2+ifelse(i3==1,0,-1)", value_cols = c("a", "r"))
}

}
