% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcalc.R
\name{calculate_formula}
\alias{calculate_formula}
\title{Calculate a measures aggregation formula on a dataset}
\usage{
calculate_formula(.data, f, id_cols = NULL, value_cols = NULL, na.rm = FALSE)
}
\arguments{
\item{.data}{Dataset (data.frame), id column requred}

\item{f}{A formula}

\item{id_cols}{Identification columns except id (string vector or R formula, like ~y)}

\item{value_cols}{Calculation columns (string vector or R formula, like ~y).
If missing, all numeric columns not including id_cols are processed}

\item{na.rm}{TRUE to replace all NAs to zero}
}
\value{
Calculation results (data.frame containing id_cols and value_cols)
        Results are placed to value_cols, id_cols copied without changes
}
\description{
Calculate a measures aggregation formula on a dataset
}
