% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{predict_data}
\alias{predict_data}
\title{Data approximation}
\usage{
predict_data(
  .data,
  .interval,
  method = "lm",
  ...,
  merge = TRUE,
  mark = FALSE,
  vars = NULL
)
}
\arguments{
\item{.data}{Source data (data.frame). Must contain interval and observation variables (for example, year and value)}

\item{.interval}{Target interval variable (vector)}

\item{method}{Approximation function ('lm', 'loess', 'glm', 'gam')}

\item{...}{Parameters for approximation function}

\item{merge}{If TRUE, results are merged with source data}

\item{mark}{If TRUE, a 'forecast' variable will be added to result}

\item{vars}{Names of interval and observation variable (1st and 2nd numeric columns by default)}
}
\value{
Approximation results (data.frame)
}
\description{
This function builds an approximation model on data containing some observation interval
and then use this model to appoximate another observation interval.
}
\examples{
\dontrun{
predict_data(.data = data.frame(
               y = c(2010, 2013, 2014),
               total = c(30800000.00, 32285714.00, 41500000.00)),
             .interval = seq(from = 2015, to = 2020, by = 1),
             method = "gam")
}

}
