% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmout.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Statistical data analysis support functions
Sergey "kol" Kolchin, 2019}
\usage{
remove_outliers(.data, vars = NULL, approx = FALSE)
}
\arguments{
\item{.data}{dataset (data.frame)}

\item{vars}{variables to inspect (character string, formula or vector).
If omitted, all numeric columns are processed. If a single variable is provided
and approx is TRUE, outlier values are removed from dataset. If it is a vector,
data is not removed but replaced with NAs.}

\item{approx}{If TRUE, outliers are approximated, if FALSE (default) - removed or replaced by NAs.
Currently not supported, reserved for further use}
}
\value{
resulting data.frame or a vector if source dataset contain only one column
}
\description{
remove_outliers function
}
\details{
Remove outliers

The function takes a dataset, checks that given variables for outliers and
either removes them or approximate using other data in these variables.
}
\examples{
\dontrun{
d <- data.frame(
  id = c("1","2","3","4","5"),
  a = c(1,2,10,1000,100),
  b = c(19999, 30000000, 1, 5, 426798))
remove_outliers(d, ~a)
remove_outliers(d)
}

}
