% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcalc.R
\name{calculate_formula}
\alias{calculate_formula}
\title{Calculate a measures aggregation formula on a dataset}
\usage{
calculate_formula(.data, f, id_cols = NULL, value_cols = NULL, na.rm = FALSE)
}
\arguments{
\item{.data}{Dataset (data.frame), id column requred}

\item{f}{A formula}

\item{id_cols}{Identification columns except id (string vector or R formula, like ~y)}

\item{value_cols}{Calculation columns (string vector or R formula, like ~y).
If missing, all numeric columns not including id_cols are processed}

\item{na.rm}{TRUE to replace all NAs to zero}
}
\value{
Calculation results (data.frame containing id_cols and value_cols)
        Results are placed to value_cols, id_cols copied without changes
}
\description{
Calculate a measures aggregation formula on a dataset
}
\examples{
\dontrun{
d <- data.frame(id = c("i1", "i1", "i2", "i3"), y = c(10, 20, 10, 10), a = c(4,2,3, 1), r = c(4,5,6,7))
calculate_formula(d, f = "log(i1)+i2+i3", id_cols = ~y, value_cols = c("a", "r"), na.rm = TRUE)

d <- data.frame(id = c("i1", "i2", "i3"), a = c(4, 3, 1), r = c(4, 6,7))
calculate_formula(d, f = "i1+i2+ifelse(i3==1,0,-1)", value_cols = c("a", "r"))
}
}
